feat(*): added a feature flag for risk overview backoffice section
