+++ apps/backoffice-v2/src/domains/workflow-definitions/fetchers.ts
@@ -25,6 +25,8 @@ export const WorkflowDefinitionConfigSchema = z
     enableManualCreation: z.boolean().default(false),
     isManualCreation: z.boolean().default(false),
     isAssociatedCompanyKybEnabled: z.boolean().default(false),
+    isCaseOverviewEnabled: z.boolean().default(false),
+    isCaseRiskOverviewEnabled: z.boolean().default(false),
     theme: WorkflowDefinitionConfigThemeSchema.default({
       type: WorkflowDefinitionConfigThemeEnum.KYB,
     }),
