+++ services/workflows-service/src/workflow/schemas/zod-schemas.ts
@@ -6,6 +6,7 @@ export const ConfigSchema = z
   .object({
     isAssociatedCompanyKybEnabled: z.boolean().optional(),
     isCaseOverviewEnabled: z.boolean().optional(),
+    isCaseRiskOverviewEnabled: z.boolean().optional(),
     isLegacyReject: z.boolean().optional(),
     isLockedDocumentCategoryAndType: z.boolean().optional(),
     isManualCreation: z.boolean().optional(),
